requirements.txt
# ExamGPT/requirements.txt
openai>=1.40.0
pillow>=10.3.0
mss>=9.0.1
opencv-python>=4.9.0.80
numpy>=1.26.4
easyocr>=1.7.1
pytesseract>=0.3.10
pyyaml>=6.0.1
pynput>=1.7.6
pyperclip>=1.8.2
pywin32>=306; platform_system=="Windows"
sympy>=1.12
